@alanpaxton 2023-10-02
I wonder if @ajkr or @pdillinger could help me here ? When I call the optimized version of `db->MultiGet()` with `verify_checksums` set (https://github.com/facebook/rocksdb/blob/main/include/rocksdb/db.h#L691) it appears that `BLOCK_CHECKSUM_COMPUTE_COUNT` is not updated, while if I call  (https://github.com/facebook/rocksdb/blob/main/include/rocksdb/db.h#L662) `BLOCK_CHECKSUM_COMPUTE_COUNT` is not updated.

I have been using the changed  `BLOCK_CHECKSUM_COMPUTE_COUNT` to confirm that the RocksJNI version of the flag is respected. My question is whether RocksDB is still behaving correctly (verifying the checksum), and the issue is that I shouldn't rely on the checksum count to check this ?